[oe][ds][ozone][input] Set the maximum input length to VKB
:Release Notes:
Implemented a way for setting the maximum input length to VKB

:Detailed Notes:
For VKB UI feature:
The Enter button will be disabled if the input field
entered the max allowed characters.
The app will set this max length of the input field and
the imemanger expected to receive from
inputMethodHost()->maxTextLength().
The Enter button will be enabled if the character length is
less than Max input length defined by the apps.

For the feature, in the patch, we get the maximum length
from HTML element and set it to VKB
via text model Wayland interface.

:Testing Performed:
RSPi4 webosose build.
Check Wayland log files and check that
max input length value is set to VKB via text model Wayland interface.

:QA Notes:
Test-cases: QWA-1847 QSDK-5201

:Issues Addressed:
[PLAT-117219] [OSE] Port [oe][ds][ozone][input] Set the maximum input length to VKB

Change-Id: I478b6926dd7add1a985f2ce3296371f0d53f3e66
Originally-Reviewed-on: http://gpro.lge.com/268667/
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/294883
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Yuri Sorochkin <yuri.sorochkin@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Sergey Kipet <sergey.kipet@lge.com>
Reviewed-by: Alexander Trofimov <alexander.trofimov@lge.com>
Reviewed-by: Dmitry Ketov <dmitry.ketov@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Yuri Sorochkin <yuri.sorochkin@lge.com>
Tested-by: Test BTS <test.bts@lge.com>