[op][upsable][file_access][whitelisting] Do not crash committing error page for blocked file access
:Release Notes:
Do not crash committing error page for blocked file access

:Detailed Notes:
RenderFrameHostImpl validation of commit params fails to realize our
error page for failing to access a local file is such. This is because
we are not using error page isolation, and the error code we get is
-10 (ACCESS_DENIED). It was only realizing it was an error if it
was 20 (BLOCKED_BY_CLIENT).

So we add access denied code too to avoid crash and show properly the
error page.

:Testing Performed:
1. Launch enyo2sampler or user guide
2. On web inspector, do window.location="/etc" or window.location="/etc/hosts"
3. It should show an error and it should not crash.

:QA Notes:
Test-cases: GT-18780, GT-18781

:Issues Addressed:
[PLAT-100581] [chromium79] File WhiteListing feature broken on chr79

Change-Id: Ib00ae0a52a744869098444e744d32c6cf828e9c7
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/269190
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/295814
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296143
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>