[op][n_upsable][watchdog] Watchdog writes callstack of stuck thread at PmLog
:Release Notes:
Watchdog writes the callstack of stuck thread at PmLog

:Detailed Notes:
When Watchdog kills process due to stuck, the callstack is retrieved
at crash report.
However, sometimes this callstack is not for the actual stuck thread.
To remain the callstack of the actual stuck thread, watchdog sends
kernel signal to stuck thread and retrieve the callstack by itself
and writes it to PmLog.

:Testing Performed:
1. generate stuck
2. wait until watchdog recovers the stuck process
3. check there is a callstack of stuck thread in PmLog

:QA Notes:
Test-cases: GT-27299

:Issues Addressed:
[PLAT-117317] Port : Watch dog

Change-Id: I33ed6ddb78c138d6213f5f0bff12b63a1b7174dc
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/267600
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/287934
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>