fixup! [op][n_upsable][media][gav] Implemented ForeignVideoWindowProvider
:Release Notes:
Fixed unwanted clip video with window bounds when switching fullscreen and
overlay mode.

:Detailed Notes:
When we update the video position we use set_crop_region with window
bounds to support part of video geometry locating outside of display
since the outside of display geometry is not supported when we are
using punch-hole.
The problem is happened in which case of when window size is changing
since setting size to blink side and wayland window is not synchronized.
This can cause unwanted clip video window by owner window bounds if
video window is set from blink before completed setting wayland window size.
Since clipping is for work-around punch-hole mode limitation with screen
bounds it is enough to clip with screen rect instead of window rect to
prevent unwanted clipping.

:Testing Performed:
Steps To Reproduce:
1. Connect screen mirroring with iphone
2. Launch AppleTV app in iphone and play any movie
3. Press blue button in MRCU
(app is changed to overlay mode)
4. Click 'Size' button on overlay airplay
5. Click 'Size' button on overlay airplay again

Expected Result:
video size is changed

Actual Result:
video size is not changed

:QA Notes:
Test-cases: GT-32685

:Issues Addressed:
[PLAT-121024] Migrate Implemented ForeignVideoWindowProvider to chromium84 OSE repo

Change-Id: Ie9717e6fd7335bff0c53d0cb0cdfe2ed01ad862d
Originally-Reviewed-on: http://gpro.lge.com/281044
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/289534
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>