[oe][ds][media][ls2] Match URIs array with URITypes
:Release Notes:
Match URIs array with URITypes

:Detailed Notes:
luna_service_uris array is indexed based on URIType, but there was
a mismatch between the sizes of each.
As some services are not available in OSE, use empty strings so that
there's no buffer overrun when accessing it based on any URIType.

:Testing Performed:
Verified that Media events from bluetooth headset are received from
MCS and the same are processed by the player. This ensures that the
service URI used to register with MCS is proper.

:QA Notes:
Test-cases: QO-1043

:Issues Addressed:
[PLAT-117435] Match URIs array with URITypes

Change-Id: I43bbcca9210dccd639f0375442178b07d6ea6230
Originally-Reviewed-on: http://gpro.lge.com/267233
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/290618
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Girish Kumar M <girish.kumarm@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Girish Kumar M <girish.kumarm@lge.com>
Tested-by: Test BTS <test.bts@lge.com>