[op][n_upsable][mse][audio] Fix pause issue for audio only playback
:Release Notes:
Fix mse playback pause does not work for audio only playback

:Detailed Notes:
Currenlty, in SetPlaybackRate of MPA, the new playback rate is
changed to 1.0 if it is audio only case. But this logic should be
useful only if the playback rate is received as 0.0f at the very
beginning of the play.
Originally this workaround was done for fixing play issue in
https://gaana.com. But now without this workaround also songs
play fine in https://gaana.com also.
So, with this patch set removing the work around. This fixes the
reported issue for cerence music and gaana.com works fine.

:Testing Performed:
1. Install the cerence music app attached in PLAT-111789
2. Launch the app and start playing a track.
3. Press pause/play button on the screen.
4. Audio should stop/start playing immediately.
5. Launch Enact web browser and load gaana.com
6. Play any music track and do play/pause from UI.
7. Audio should start/stop immediately

:QA Notes:
Test-cases: QWA-4495

:Issues Addressed:
[PLAT-121266] Fix pause issue for audio only playback

Change-Id: Idd72f5bfca5e262de4e599461eeefa6adf75d786
Originally-Reviewed-on: http://gpro.lge.com/282753
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/291624
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Girish Kumar M <girish.kumarm@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>