fixup! [op][n_upsable][media][gav] Implemented ForeignVideoWindowProvider
:Release Notes:
In webos, for application window resolution is less than the screen
resolution, we have to consider window bounds to decide full-screen
mode of video.

:Detailed Notes:
Sometimes application window resolution can be less than the screen
resolution. In webos, always the application is displayed in
full screen mode. When video is set to full screen by setting its width,
height to window.innerWidth and window.innerHeight, then comparing the
dest rect with screen rect, its calculated as non full screen and
the video is scaled to screen rect. But it should be set to window rect,
because the application window is already scaled up and its displaying
in full screen mode. so we loose the aspect ratio of video in full
screen mode in this case. With this change we use window rect instead of
screen rect to decide full screen. However we retain to use screen rect
for calculating visible_rect

:Testing Performed:
1. In appinfo.json, set "resolution": "1280x720".
2. Set video's width to window.innerWidth and video's height to
   window.innerHeight for full screen mode.
3. Observe video plays with original width/height ratio instead of
   screen resolution.

:QA Notes:
Test-cases: GT-28432

:Issues Addressed:
[PLAT-123536] Migrate ForeignVideoWindowProvider fixup patch to OSE

Change-Id: Id7dfea19a3e7e9b40bdecf78181c3c5349bf3d9d
Originally-Reviewed-on: http://gpro.lge.com/296138
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/295024
Tested-by: Girish Kumar M <girish.kumarm@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Girish Kumar M <girish.kumarm@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>