[op][upsable][scroll] Clamp scroll offsets passed to CC from impl side
:Release Notes:
Clamp scroll offsets passed to CC from impl side.

:Detailed Notes:
There is a scenario where, on scrolling fast and close to the limits of
the scrollable area bounds, it could end up generating scroll offsets
out of the scrollable bounds (negative values, or bigger than maximum
scroll offset).

Once that happens, native scroll starts to fail, as it cannot handle
scrolls offset out of the bounds. So, in case the scrollable contents
would be assigned to the layer with wrong scroll_offset (or any child of
it), it ends up failing to scroll, and CC may be confused and draw the
scrolled layer at an unexpected position.

In the detected issue, the scenario was visible when introducing
will-change: transform (or an actual 2d transformation) as it would
make the items be in the scrollable layer.

This scenario was independent of using webOS native scroll (though it
is easier to reproduce with it as scrolling is faster).

It is not clear where this situation is introduced. But we can avoid
the problem just clamping the value when it is set from impl side.

:Testing Performed:
1. Plug an USB disk with a number of photos that is not very big in top
  folder (20-30 is OK).
2. Open Photo&Video
3. Go to USB disk
4. In web inspector, select the node that has the voice intent "Scroll"
   and is parent of the list of items in screen. Add to that node the
   CSS line "will-change: transform;"
5. Scroll up and down to the top repeatedly with remote wheel.

:QA Notes:
Test-cases: GT-27823

:Issues Addressed:
[PLAT-85038] will-change:transform cause incorrect layout when scrolling item in music app list
[PLAT-118555] Port : [op][upsable][scroll] Clamp scroll offsets passed to CC from impl side

Change-Id: I37c9a29e7cdb7d4fc95dca2703802a6793e1defc
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/268473
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/295972
Tested-by: Jihye Hong <jh.hong@lge.com>
Reviewed-by: Jihye Hong <jh.hong@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296139
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>