[op][ds][cbe] Do not flag cbe or app_shell as testonly
:Release Notes:
Do not flag cbe or app_shell as test only.

:Detailed Notes:
When using enactbrowser, app_shell is our production browser, so it should
not have any test library or component. So we need to modify the original
app_shell testonly flags to be unset. Code needs to be modified so a lighter
version of content_shell_lib (named conntent_shell_nontest_lib) can be
used from app_shell.

:Testing Performed:
Build. Use strings on libcbe.so to check ShellBrowserMain is not present.

:QA Notes:
Test-cases: GT-26820

This is not reducing size back to 70M as we had in chromium68. Mojoification
has made Chromium bigger in general. Though, with this change we drop around
150K that is the implementation of the content shell test components.

In chromium79 libcbe.so size is around 80MB after applying this.

:Issues Addressed:
[PLAT-81235] [chromium72] libcbe.so disk size is bigger than 70MB
[PLAT-117515] Port : [op][ds][cbe] Do not flag cbe or app_shell as testonly

Change-Id: Iea13797aec54c2d87819da8283b579473cfe8d20
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/267812
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/296375
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296814