[op][n_upsable][memory] Set the discardable memory limit based on size of shared memory
:Release Notes:
Set the discardable memory limit based on size of shared memory

:Detailed Notes:
The runtime flag value "shared-mem-system-mem-reduction-factor" is
calculating the maximum limit of discardable memory of Chromium.
But current logic is calculating based on the system memory not the
shard memory, i.e., limit = total_system_memory / reduction_factor.

This is wrong and have problems on devices which have large system
memory like webOS Auto.
In webOS Auto, the system memory is 8GB and the reduction factor is 10
so the limit will be (8GB / 10) = 800MB but the default total size of the
shared memory is around 400MB and cause a crash when Chromium is
allocating discardable memory more then shared memory limit.

This patch changes to calculate the reduction factor based on size of
shared memory rather than the system memory.

:Testing Performed:
1) Launch YouTube test app
2) Scroll the page way down until can't scroll anymore
3) Click the YouTube logo on top left and refresh the page
4) Repeat step 2~3 several times
5) Verify that there is no crash(no instant re-launch of the app)

:QA Notes:
Test-cases: GT-37485

:Issues Addressed:
[PLAT-113301] Youtube app gets freezed and re-launches
[PLAT-118536] Port : [op][n_upsable][memory] Set the discardable memory limit based on size of shared memory

Change-Id: Iaee33012dff7c21380802c498f7ff198434faeb3
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/281630
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/296372
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Jihye Hong <jh.hong@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jihye Hong <jh.hong@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296935
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>