fixup! [op][n_upsable][webrtc] Implement video decoder for webrtc
:Release Notes:
Fix video conference sometime crash when do join/unjoin from
remote devices

:Detailed Notes:
During decode we check using client, whether decoder resource are
available or not and depending on the same we decide to fall back
to s/w decoder. So to get the status of decoder available or not,
we had to use lock to get the decoder instance and set the client
and use the same client instance to call HasAvailableResources.

We tried to create a sync between the decoder and the player. But
it leads to side effects and sometime random crashes during join
or unjoin of remote users.

Also when a platform resource is not available and the resource
is not acquired for the decoder yet then we have to fallback to
s/w decoder immediately.

So in this patch set we removed the use of pass through decoder
client and decoder id to create a mapping. Instead we use a global
variable as media_decoder_available_, to store the status of decoder
availability. It is set to FALSE when we get resource allocation
error from the platform media pipeline. Also we restore the status
back to TRUE when an acquired resource is freed.

Rest of the changes are derived following the upstream class used
for video decoding accelaration - RTCVideoDecoderAdapter

We also added a runtime flag "enable-webrtc-platform-video-decoder"
which allows us to enable/disable the use of h/w decoder in runtime.

:Testing Performed:
1. Tested Video conference with H264 & VP8 codec types using
   https://appr.tc/?vsc=H264 and https://appr.tc/?vsc=VP8
   in Auto board
2. Tested video conference with H264 codec type using
   https://appr.tc/?vsc=H264 in OSE board
3. Tested Video conference using https://meet.jit.si/
4. Tested streaming from canvas element using below url.
   https://webrtc.github.io/samples/src/content/capture/canvas-pc/

:QA Notes:
Test-cases: QWA-5260, QWA-2739, QWA-2743

:Issues Addressed:
[PLAT-126048] Migrate "fixup Implement video decoder for webrtc" patch

Change-Id: I511e33bfec6bf065867ff444644866e32c224cec
Originally-Reviewed-on: http://gpro.lge.com/294801
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296848
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>