[op][n_upsable][build] Add configurations to build qemux86
:Release Notes:
Add configurations to build qemux86

:Detailed Notes:
Until Chromium72, when building qemux86 we used gcc as a host machine
compiler(instead the clang as default) to fix the V8 compile error.
But from Chromium79, there were some problem with this solution when
the CPU architecture is same but the instruction length is
different(32 bit vs. 64 bit) with the target machine and the host
machine.

The proper solution would be using clang as a host machine compiler,
and this patch add configurations for that.

:Testing Performed:
Build succeed

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[PLAT-119477] Build qemux86 for OSE

Change-Id: I0ba83921a2a52ff6cec074d9319950f66250536f
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296590
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Heegoo Han <heegoo.han@lge.com>