[op][n_upsable][media] Fix passing invalid window id to platform
:Release Notes:
Fix passing the invalid native window id to platform.

:Detailed Notes:
Sometimes the value of native window id is lost because the character
pointer is getting out of scope when the thread executes the function.
Added changes to create the copy of window id using std::string and
passed the string object by value, so the compiler will create the
copy of object and the window id value is retained.

:Testing Performed:
1. Launch "Shaka Player" built-in app
2. Click ALL CONTENT menu at top side.
3. From the combo box on the DRM text below at topside, Select the
   Widevine value to find Widevine videos.
4. Play "Angel one" video multiple times.
5. It should play without error popup.

:QA Notes:
Test-cases: GT-27193

:Issues Addressed:
[PLAT-123996] Angel one video is not played in SHAKA PLAYER app

Change-Id: Ibc21ba71a74aa80d0bfe6a02136453a9ad2600a1
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/296707
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>