fixup! [op][n_upsable][splash][fmp] New splash and FMP implementation
:Release Notes:
Do not mix surface activation of different applications.

:Detailed Notes:
The m84 port of the new splash screen implementation was not properly filtering
the SurfaceActivatedEx events to DisplayDamageTracker and DisplayScheduler.
This would make new application activation happen with frames of other
applications.

With this change, surface hierarchy is matches to only acknowledge the events
for surfaces children of the root surface id of the damage tracker.

:Testing Performed:
1. restart webapp-mgr
2. Launch firstuse
3. Launch apple airplay with content url.

Now it does not mix events.

:QA Notes:
Test-cases: GT-15540, GT-18470, GT-18526, GT-23384, GT-23629, GT-23844,
    GT-26460, GT-27610, GT-27801, GT-27804, GT-27838, GT-27922, GT-27924,
    GT-28157, GT-28088, QS-8559

:Issues Addressed:
[PLAT-126662] First use app is NOT running

Change-Id: I3a777abfe9b54e1c65f1e319d2f82877b1f4ff60
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/296978
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/297144
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>