fixup! [op][n_upsable][webrtc] Implement web media player for webrtc
:Release Notes:
Multiple fixes for WebMediaPlayerWebRTC and pass through decoder
related to the use of h/w resources

:Detailed Notes:
Currently we store h/w decoder is not available in a static
variable when we get resource alloc error from pipeline . As a
result next decoders are not trying to use h/w resources and
falling back to use s/w decoder. There are many scenarios where
h/w resources might be free but we are not going to use them
because of this. So with this patch set changed the logic to
try using the h/w resource by default and in case there is
resource alloc error then fallback to s/w decoder.

Also in sites like Jitsi meet when we toggle between tile view the
player instance is destroyed and recreated for the same user. But
we may not receive the key frames because of which the pipeline is
not able to configure properly and black patch video is visible.

So, we set the key frame request when ever there is player resume
after suspend or first time when we go for handling encoded frames
ensuring to get key frames at the beginning of pipeline creation.
In this scenario we are also adjusting the timestamp value of the
encoded buffers so that the playback continues without delay.

Also reorganized some code like below
1. Added CompositorType for separating video layer creation
2. Separation functions for create/destroy media platform API
3. Recreation of MediaPlatformAPI if codec type changed
4. Removal of handle encoded frame usage
5. Removal the usage of unnecessary EnqueueHoleFrame
6. Avoiding frame resize for encoded frames

:Testing Performed:
1. Launch WebRtc test app (com.webos.app.test.webrtc)
2. load the https://meet.jit.si/ test page using side loading from
   web inspector from desktop
3. Start a meeting by entering some unique ID.
4. start the meeting from some other device say laptop/smart phone
   having a google chrome browser
5. Enter the same meeting id and join the conference.
6. Verify that the playback is smooth for minimum two remote videos
7. Also verify that more then one instance of GMP is running

:QA Notes:
Test-cases: QWA-2768, QWA-2744, QWA-5246

:Issues Addressed:
[PLAT-127889] Fix black screen issue in Jitsi Meet tile mode

Change-Id: I3b25f7c1ef29fa68ca7fef61b080b5e57a191091
Originally-Reviewed-on: http://gpro.lge.com/297903
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/299971
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Test BTS <test.bts@lge.com>