[oe][n_upsable][network][recovery] Add display affinity to network settings parameters
:Release Notes:
Add display affinity to network settings parameters to launch the settings app
on the same display as the network error page.

:Detailed Notes:
When "NETWORK SETTINGS" button in a network error page is clicked, the settings
app is launched on the first display regardless of the display of the error
page, because the settings app does not know the display affinity of the caller.
LaunchNetworkSettings of NetworkErrorPageController takes only one parameter
target_id.

To launch the settings app on the same display as the network error page,
LaunchNetworkSettings of NetworkErrorPageController is changed to take another
parameter display_affinity. The display affinity is taken from
window.webOSSystem.launchParams using JavaScript.

:Testing Performed:
1. Disconnect network and launch YouTube web app with displayAffinity set to 0.
   luna-send -n 1 -f luna://com.webos.service.applicationmanager/launch '{"id":"com.webos.app.test.youtube", "params": {"displayAffinity": 0}}'
   "CHECK NETWORK CONNECTION(-106)" page appears.
2. Click "NETWORK SETTINGS" button.
   Settings app appears on the same display as the network error page.
3. Close the Settings and YouTube apps.
4. Launch YouTube web app on another display.
   luna-send -n 1 -f luna://com.webos.service.applicationmanager/launch '{"id":"com.webos.app.test.youtube", "params": {"displayAffinity": 1}}'
   "CHECK NETWORK CONNECTION(-106)" page appears.
5. Click "NETWORK SETTINGS" button.
   Settings app appears on the same display as the network error page.

:QA Notes:
Test-cases: QWA-3391

:Issues Addressed:
[PLAT-103613] [Auto][OSE] Support multi-display on network error page
[PLAT-128851] [webOS OSE]:  Settings app is launched display 0 instead of display 1.

Change-Id: Iab29f1f4f5a4ff7d0255257f17ca1a001bdfe187
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/272972
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/298556
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>