fixup! [op][ds][media][gmp] Implement MediaPlatformAPIWebOSGmp
:Release Notes:
Fix to pass video info to natural video size change callback

:Detailed Notes:
Currently in SetMediaVideoData we try to get "video" object
from the root object. But actually we are getting "video" wrapped
in side "videoInfo" object. As a result we dont get the desired
object and natural_video_size_changed_cb_ is not invoked.

With this patch set we extract "video" from "videoInfo" object
and are able to pass the width, height info to the registered
callback.

:Testing Performed:
Verified the below steps with and without --enable-neva-media-service
in web app manager script:
1. Launch "com.webos.app.test.youtube" application
2. Play any video
3. Video playback starts
4. Verify Pause, Play and Seek functionalities
5. Press full screen button
6. Verify Pause, Play and Seek functionalities in fullscreen mode
7. Return to inline mode
8. Exit the application
9. Also, verified the mentioned test case

:QA Notes:
Test-cases: QR-363

:Issues Addressed:
[PLAT-129631] Fix video info passing to natural_video_size callback

Change-Id: I1a5e19aae6818a420dd82029642107621747d11f
Reviewed-on: http://gpro.lge.com/c/webosauto/chromium84/+/299009
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Girish Kumar M <girish.kumarm@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/299970
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>