fixup! [op][ds][splash] Defer commits while loading background images
:Release Notes:
Improve splash screen removal while background images are loaded.

:Detailed Notes:
In many cases we find that FCP and FMP are called on first layout including
images, but that does not actually wait for the images to be ready, just
for them to start to be painted. This makes the background images
deferral less useful (it will avoid intermediate paints, but it will not
hold splash screen removal).

To fix this, this change delays paint timing notifications used by splash
screen detection (as First Contentful Paint) while background images are
loaded. Once loaded they are emited as expected.

:Testing Performed:
Reboot OSE device showing bareapp on start, see if the transition shows
a grey screen. It should be gone after this patch.

:QA Notes:
Test-cases: GT-23384

:Issues Addressed:
[PLAT-128187] While rebooting, grey screen is displaying after launching bare app

Change-Id: I043f59490ed5d5eae1299ac0b8ac8ec8c01f1ab5
Reviewed-on: http://gpro.lge.com/c/webosauto/chromium84/+/302505
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Yuri Sorochkin <yuri.sorochkin@lge.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Yuri Sorochkin <yuri.sorochkin@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/302611
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>