fixup! [op][n_upsable][webrtc] Implement webrtc video encoder interface.
:Release Notes:
Use common video encoder interface for all targets in webrtc

:Detailed Notes:
Currently WebRtcVideoEncoderWebOSGMP is using the interface
provided by g-camera-pipeline and the same can not be used for
all the platforms. So common interface as VideoEncoderAPI is
introduced by webos to make chromium changes common for all
platforms.

With these changes we remove the use of media encoder client
and use the newly introduced mcil::encoder::VideoEncoderAPI
from media-codec-interface to communicate with platform provided
h/w encoder. Since this implementation is going to be same for
all targets so renamed WebRtcVideoEncoderWebOSGMP as
WebRtcVideoEncoderWebOS also.

:Testing Performed:
1. Tested Video conference with VP8, H264 codec types using
   https://appr.tc/?vsc=H264 and https://appr.tc/?vsc=VP8
   in Auto board
2. Tested video conference with H264 codec type using
   https://appr.tc/?vsc=H264 in OSE board
3. Tested Video conference using https://talky.io/
4. Tested streaming from canvas element using below url.
   https://webrtc.github.io/samples/src/content/capture/canvas-pc/
5. Streaming from video element status is same as before and it
   wont work with neva media

:QA Notes:
Test-cases: QWA-5260, QWA-2739, QWA-2743

:Issues Addressed:
[PLAT-121123] Implement h/w encoding framework in Chromium.

Change-Id: I0e9a4216d1f7405152a3ae13f849a39e2b2a3a0d
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/304598
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Karthick Sundaram <karthick.sundaram@lge.com>
Tested-by: Test BTS <test.bts@lge.com>