fixup! [op][n_upsable][webrtc] Implement webrtc video encoder interface
:Release Notes:
Fix glitches while streaming from canvas element to peer connection

:Detailed Notes:
When we stream from canvas element, the buffer we receive from
canvas is different compared to the buffers we receive from camera.
As a result buffer extraction for YUV data is failing in
media-codec-interface similar to OSE and AUTO when we use GST media.
With this patch set, we are using webrtc::ExtractBuffer for non GST
media implementation also. So that Y, U, V data can be extracted and
passed correctly to media-codec-interface.
This also fixes sometime crash due to wrong buffer pointer extraction.

:Testing Performed:
1. Load "Stream from canvas element" test page using webrtc test app
   https://webrtc.github.io/samples/src/content/capture/canvas-pc/
2. Move the tea-pot on left side using mouse/touch.
3. Teapot Image on the right side box should move accordingly.

:QA Notes:
Test-cases: GT-43084

:Issues Addressed:
[PLAT-137012] WebAppMgr crash is displayed in WebRTC Test page.

Change-Id: Ib1749c970d4b8d7c730073c437018e229cdb057b
Originally-Reviewed-on: http://gpro.lge.com/305580
Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/305581
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Karthick Sundaram <karthick.sundaram@lge.com>
Tested-by: Test BTS <test.bts@lge.com>